Restore deletion of the test results directory
By default, the Git build process deletes the test results directory, at
least if it hasn't hit any failures.  That was previously disabled
because I thought that was the only place that kept a record of the
build results, but I've now found that information is in a more
accessible format in cygport's log directory, so the patch is redundant.