Use native Cygwin installer rather than Choco
The Chocolatey-based installer for Cgywin works, but it's slow, as it
installs each package separately, which means a separate run of all the
Cygwin post-install scripts.  Instead, use a GitHub action that just
calls through to the standar Cygwin installer, which is quicker, and is
also actually properly supported by the Cygwin project.