Set the correct path for AsciiDoc dblatex files
It looks like this changed some time around AsciiDoc v10, and the files
are now stored in the same path as the rest of the Python library files.
As part of the build process, get the current location of the files, and
specify that as part of the make options.