Remove dependency on GitHub runner layout
GitHub runners are single-user and disposable, so it's safe to assume --
regardless of the recorded owner of any filesystem object -- that any
Git directories are safe.  That means there's no need for the safe
directory check, so opt out of it, removing the reliance on the specific
filesystem layout of the GitHub runners.