Bump CI to use actions/checkout@v3
This means we can tidy up some of the workarounds for the new Git
safe.directory config handling.  It's still necessary to handle that
manually, as actions/checkout will otherwise assume the location of the
Git config file, and get it wrong because it's expecting Git for
Windows, not Cygwin Git.