Add gettext to Git base dependencies
Git will mostly work without gettext, but it will get very upset if
there's a non-Cygwin gettext in PATH, and it's being built expecting
gettext to be present at runtime.  Because gettext isn't a library,
cygport doesn't have a chance of finding the dependency automatically,
so just specify it manually.