Split t7508 out
I think this is another space-hungry test, which is causing the overall
test runs to fail.  Try a test where that has been separated out.

Don't test for how much disk space is available before running the space
hungry tests -- we can just let them fail if they're going to fail --
but do report how much disk space is available to stdout so if there is
still a problem it's a bit more obvious.